[SPARK-27687][SS] Rename Kafka consumer cache capacity and document caching
